new combatant detail takes its rightful place
